Handle empty blocklist and missing values
The blocklist may not always have elements in them, or may have elements with
missing fields. For example, the "value" field may be repeated incorrectly or
may be missing altogether. In those cases, treat the corresponding blocklist
as missing and handle the "vacancy" appropriately, ie, do not treat the vacancy
as a hit in the cache, for example, or even as a valid "blocklist value".